Version 6.7 Community Edition Update

+++ src/services/ddsi2/code/q_bswap_template.c
@@ -1,3 +1,22 @@
+/*
+ *                         OpenSplice DDS
+ *
+ *   This software and documentation are Copyright 2006 to TO_YEAR PrismTech
+ *   Limited, its affiliated companies and licensors. All rights reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
 /* -*- c -*- */
 
 #if defined SUPPRESS_BSWAP_INLINES && defined NN_C99_INLINE
@@ -7,7 +26,7 @@
 
 NN_C99_INLINE unsigned short bswap2u (unsigned short x)
 {
-  return (x >> 8) | (x << 8);
+  return (unsigned short) ((x >> 8) | (x << 8));
 }
 
 NN_C99_INLINE unsigned bswap4u (unsigned x)
